Moved model.ts
And fixed typing errors.
